Resources to long

+++ WarOfEmpires.Domain/Common/Resources.cs
@@ -3,16 +3,16 @@ using System.Collections.Generic;
 
 namespace WarOfEmpires.Domain.Common {
     public sealed class Resources : ValueObject {
-        public int Gold { get; private set; }
-        public int Food { get; private set; }
-        public int Wood { get; private set; }
-        public int Stone { get; private set; }
-        public int Ore { get; private set; }
+        public long Gold { get; private set; }
+        public long Food { get; private set; }
+        public long Wood { get; private set; }
+        public long Stone { get; private set; }
+        public long Ore { get; private set; }
 
         private Resources() {
         }
 
-        public Resources(int gold = 0, int food = 0, int wood = 0, int stone = 0, int ore = 0) {
+        public Resources(long gold = 0, long food = 0, long wood = 0, long stone = 0, long ore = 0) {
             if (gold < 0) throw new ArgumentOutOfRangeException("gold", "Negative values are not allowed");
             if (food < 0) throw new ArgumentOutOfRangeException("food", "Negative values are not allowed");
             if (wood < 0) throw new ArgumentOutOfRangeException("wood", "Negative values are not allowed");
@@ -68,16 +68,16 @@ namespace WarOfEmpires.Domain.Common {
             return new Resources(a.Gold - b.Gold, a.Food - b.Food, a.Wood - b.Wood, a.Stone - b.Stone, a.Ore - b.Ore);
         }
 
-        public static Resources operator *(Resources a, int b) {
+        public static Resources operator *(Resources a, long b) {
             return new Resources(a.Gold * b, a.Food * b, a.Wood * b, a.Stone * b, a.Ore * b);
         }
 
-        public static Resources operator *(int a, Resources b) {
+        public static Resources operator *(long a, Resources b) {
             return b * a;
         }
 
         public static Resources operator *(Resources a, decimal b) {
-            return new Resources((int)(a.Gold * b), (int)(a.Food * b), (int)(a.Wood * b), (int)(a.Stone * b), (int)(a.Ore * b));
+            return new Resources((long)(a.Gold * b), (long)(a.Food * b), (long)(a.Wood * b), (long)(a.Stone * b), (long)(a.Ore * b));
         }
 
         public static Resources operator *(decimal a, Resources b) {
